+++ ipbb/tools/common.py
@@ -71,3 +71,29 @@ class SmartOpen( object ):
     self.file.write( ' '.join( strings ) )
     self.file.write( "\n" )
 #------------------------------------------------------------------------------
+
+
+#------------------------------------------------------------------------------
+class AAA(object):
+    def __init__(self, prefix=None):
+        self._write = sys.stdout.write
+        self._flush = sys.stdout.flush
+        self.prefix = prefix
+
+    def __del__(self):
+        # self.close()
+        pass
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, *args):
+        # self.close()
+        pass
+
+    def write(self, message):
+        self._write(message.replace('\n','\n'+self.prefix) if self.prefix else message)
+
+    def flush(self):
+        self._flush()
+#------------------------------------------------------------------------------
